+++ eventium-sql-common/src/Eventium/Store/Sql/DefaultEntity.hs
@@ -21,8 +21,6 @@ module Eventium.Store.Sql.DefaultEntity
   )
 where
 
-import Data.Text (Text)
-import Data.Time (UTCTime)
 import Database.Persist (Key)
 import Database.Persist.Sql (fromSqlKey, toSqlKey)
 import Database.Persist.TH
@@ -38,11 +36,8 @@ share
 SqlEvent sql=events
     uuid UUID
     version EventVersion
-    eventType Text
-    event JSONString
-    correlationId UUID Maybe
-    causationId UUID Maybe
-    createdAt UTCTime Maybe
+    payload JSONString
+    metadata JSONString Maybe
     UniqueUuidVersion uuid version
     deriving Show
 |]
@@ -63,13 +58,10 @@ defaultSqlEventStoreConfig =
       sqlEventStoreConfigUnKey = sqlEventUnKey,
       sqlEventStoreConfigUUID = sqlEventUuid,
       sqlEventStoreConfigVersion = sqlEventVersion,
-      sqlEventStoreConfigEventType = sqlEventEventType,
-      sqlEventStoreConfigData = sqlEventEvent,
-      sqlEventStoreConfigCorrelationId = sqlEventCorrelationId,
-      sqlEventStoreConfigCausationId = sqlEventCausationId,
-      sqlEventStoreConfigCreatedAt = sqlEventCreatedAt,
+      sqlEventStoreConfigData = sqlEventPayload,
+      sqlEventStoreConfigMetadata = sqlEventMetadata,
       sqlEventStoreConfigSequenceNumberField = SqlEventId,
       sqlEventStoreConfigUUIDField = SqlEventUuid,
       sqlEventStoreConfigVersionField = SqlEventVersion,
-      sqlEventStoreConfigDataField = SqlEventEvent
+      sqlEventStoreConfigDataField = SqlEventPayload
     }