+++ examples/cafe/src/Cafe/ChefTodoList.hs
@@ -38,7 +38,7 @@ chefTodoListMain = do
       globalReader = runEventStoreReaderUsing (`runSqlPool` pool) cliGloballyOrderedEventStore
       sub = pollingSubscription globalReader checkpoint 1000
       handler = EventHandler $ \globalEvent -> do
-        let inner = streamEventEvent globalEvent
+        let inner = streamEventPayload globalEvent
         case traverse (decode jsonStringCodec) inner of
           Nothing -> return ()
           Just tabEvent -> do
@@ -54,7 +54,7 @@ handleEventToMap foodMap (StreamEvent uuid _ _ (TabClosed _)) = Map.delete uuid 
 handleEventToMap foodMap streamEvent =
   let uuid = streamEventKey streamEvent
       oldList = Map.findWithDefault [] uuid foodMap
-   in Map.insert uuid (handleEventToFood oldList $ streamEventEvent streamEvent) foodMap
+   in Map.insert uuid (handleEventToFood oldList $ streamEventPayload streamEvent) foodMap
 
 handleEventToFood :: [Maybe Food] -> TabEvent -> [Maybe Food]
 handleEventToFood oldFood (FoodOrdered newFood) = oldFood ++ map Just newFood