+++ flake.nix
@@ -6,14 +6,20 @@
     flake-utils.url = "github:numtide/flake-utils";
   };
 
-  outputs = { self, nixpkgs, flake-utils }:
-    flake-utils.lib.eachDefaultSystem (system:
+  outputs =
+    {
+      self,
+      nixpkgs,
+      flake-utils,
+    }:
+    flake-utils.lib.eachDefaultSystem (
+      system:
       let
         pkgs = nixpkgs.legacyPackages.${system};
-        
+
         # Use GHC 9.6.7 as requested
         hPkgs = pkgs.haskell.packages.ghc967;
-        
+
         # Define all the local packages
         localPackages = [
           "eventium-core"
@@ -28,14 +34,16 @@
         ];
 
         # Function to build a single package
-        buildPackage = name: 
-          hPkgs.callCabal2nix (builtins.replaceStrings ["/"] ["-"] name) ./${name} {};
+        buildPackage =
+          name: hPkgs.callCabal2nix (builtins.replaceStrings [ "/" ] [ "-" ] name) ./${name} { };
 
         # Build all packages
-        packages = builtins.listToAttrs (map (name: {
-          name = builtins.replaceStrings ["/"] ["-"] name;
-          value = buildPackage name;
-        }) localPackages);
+        packages = builtins.listToAttrs (
+          map (name: {
+            name = builtins.replaceStrings [ "/" ] [ "-" ] name;
+            value = buildPackage name;
+          }) localPackages
+        );
 
         # Development dependencies
         devDependencies = with pkgs; [
@@ -44,12 +52,12 @@
           hPkgs.cabal-install
           hPkgs.hpack
           stack
-          
+
           # Database libraries/headers for building (no local server needed)
           postgresql.lib
           postgresql.dev
           sqlite
-          
+
           # Development tools
           hPkgs.haskell-language-server
           hPkgs.hlint
@@ -57,17 +65,18 @@
           hPkgs.ghcid
           hPkgs.hspec-discover
           just
-          
+
           # System dependencies
           pkg-config
           zlib
         ];
 
-      in {
+      in
+      {
         # Export all packages
         packages = packages // {
           default = packages.eventium-core;
-          
+
           # Combined package that builds everything
           all = hPkgs.buildStackProject {
             name = "eventium-all";
@@ -85,7 +94,7 @@
         # Development shell
         devShells.default = pkgs.mkShell {
           buildInputs = devDependencies;
-          
+
           shellHook = ''
             just hpack
             echo ""
@@ -105,16 +114,17 @@
           default = flake-utils.lib.mkApp {
             drv = packages.eventium-core;
           };
-          
+
           # Example apps
           bank-example = flake-utils.lib.mkApp {
             drv = packages.examples-bank;
           };
-          
+
           counter-cli = flake-utils.lib.mkApp {
             drv = packages.examples-counter-cli;
           };
         };
+      }
+    );
+}
 
-      });
-}